bazaar-updater/requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
PyYAML==6.0.2
requests==2.29.0
urllib3==1.26.20
